airai-api/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
exceptiongroup==1.1.1
fastapi==0.95.0
frozenlist==1.3.3
geojson==2.5.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
langchain==0.0.133
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.0
pluggy==1.0.0
pydantic==1.10.7
pyowm==3.3.0
PySocks==1.7.1
pytest==7.2.2
pytest-asyncio==0.21.0
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
tenacity==8.2.2
tomli==2.0.1
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
yarl==1.8.2
slowapi~=0.1.8